% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbg.R
\docType{data}
\name{colores_nsdl4tb}
\alias{colores_nsdl4tb}
\title{Funciones para graficar}
\format{An object of class \code{character} of length 56.}
\usage{
colores_nsdl4tb
}
\description{
Colores
}
\keyword{datasets}

