% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbg.R
\name{tbg_hc_fisico_unit}
\alias{tbg_hc_fisico_unit}
\title{Funciones para graficar}
\usage{
tbg_hc_fisico_unit(data, tipo)
}
\arguments{
\item{data}{datos tidy de tmb}

\item{tipo}{"line" o "column"}
}
\description{
Grafica de inventario físico
}

