% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbg.R
\name{tbg_hc_pagos}
\alias{tbg_hc_pagos}
\title{Funciones para graficar}
\usage{
tbg_hc_pagos(data, tipo)
}
\arguments{
\item{data}{datos tidy}

\item{tipo}{"line" o "column"}
}
\description{
Grafica de datos tidy
}

