% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi_h_escogerlocal}
\alias{tbi_h_escogerlocal}
\title{Funciones para inicializar}
\usage{
tbi_h_escogerlocal(l)
}
\arguments{
\item{l}{Local (Computadora): 1 = Negocio local, 2 = Casa (Personal), 3 = Usuario}
}
\description{
Retorna string de path a archivos, dependiendo el local
}

