% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi_h_renombrar}
\alias{tbi_h_renombrar}
\title{Funciones para inicializar}
\usage{
tbi_h_renombrar(df, columnas, tipo = "pre", char)
}
\arguments{
\item{df}{data.frame}

\item{columnas}{Numeros de columnas a renombrar}

\item{tipo}{tipo de renombre ("pre" = antes o "pos" = despues)}

\item{char}{caracteres a poner antes o después}
}
\description{
Renombra una serie de columnas, manteniendo el nombre original
}

