% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi_importar}
\alias{tbi_importar}
\title{Funciones para inicializar}
\usage{
tbi_importar(db, l)
}
\arguments{
\item{db}{base de datos ("cortes")}

\item{l}{Local (Computadora): 1 = Negocio local, 2 = Casa (Personal), 3 = Usuario}
}
\description{
Importa y guarda en data.frame objetos de diferentes bases de datos.
}

