% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi_importer_cortes}
\alias{tbi_importer_cortes}
\title{Funciones para inicializar}
\usage{
tbi_importer_cortes(archivo)
}
\arguments{
\item{archivo}{archivo de cortes}
}
\description{
Importar y unir cortes diarios
}

