% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi_importer_seps}
\alias{tbi_importer_seps}
\title{Funciones para inicializar}
\usage{
tbi_importer_seps(archivo, inventarios, ticks, logs, cat_metales, cat_fam)
}
\arguments{
\item{archivo}{archivo de inventario}

\item{cat_metales}{archivo de catalogo de metales}

\item{cat_fam}{archivo de catalogo de familias}
}
\description{
Importar y unir, con catalogo, ventas
}

