% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbm.R
\name{tbm_crear_graficar}
\alias{tbm_crear_graficar}
\title{Funciones para manipular}
\usage{
tbm_crear_graficar(data, tipo)
}
\arguments{
\item{data}{Datos}

\item{tipo}{Actualmente por: "pagos"}
}
\description{
ojo, cambiar el nombre!
Manipula datos de corte, produce tipo de objeto tidy para gráficar
}

