% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbm.R
\name{tbm_grafica_fisicos}
\alias{tbm_grafica_fisicos}
\title{Funciones para manipular}
\usage{
tbm_grafica_fisicos(data)
}
\arguments{
\item{data}{Datos}
}
\description{
Manipula datos de inventario fisico para gráficar
}

