% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbm.R
\name{tbm_limpiar_ventas}
\alias{tbm_limpiar_ventas}
\title{Funciones para manipular}
\usage{
tbm_limpiar_ventas(venta)
}
\arguments{
\item{venta}{Datos de venta (no headers) de tbi}
}
\description{
Manipula datos de ventas para gráficar
}

