% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbm.R
\name{tbm_modelar_ingreso}
\alias{tbm_modelar_ingreso}
\title{Funciones para manipular}
\usage{
tbm_modelar_ingreso(cortes, venta)
}
\arguments{
\item{cortes}{tabla de cortes}

\item{venta}{venta (header)}
}
\description{
Une cortes con ventas y obtiene diferencias atribuibles a otros tipos de ingresos.
}

